Support building hook as Client of existing PAYDAY 2 hook.
This adds a number of preprocessor if(n)defs that match on a
compile-time var. If the var is set, the code is altered to build as a
guest of the existing PAYDAY 2 Lua hook that has a C API. After
compilation the library should be placed in lib/Native per the
documentation.

This commit does not alter the project configuration. If you wish to
make a build configuration using this functionality, create a copy of
the Release configuration and then edit it to set BLT_AS_CLIENT for the
preprocessor.