Add BLT runtime detection support for client/standalone mode.
The BLT hook will determine at runtime whether it is running under the
existing hook and behave appropriately according to the situation.

Namely; if running as a standalone, signature lookups will be performed
in order to populate function pointers. Otherwise, it will simply use
GetProcAddress to fetch function addresses from the existing hook.