Regretfully restore lua_call() crash prevention.
There's a variety of good reasons for crashing on a bad lua_call(),
unfortunately, it looks like writing reliable code isn't much of a
priority for GoonMod.

This commit causes lua_call() to always be detoured with a function that
instead invoked lua_pcall(), with all the caveats that implies.